@@ -339,8 +339,8 @@ public class Attributes {
     @Override
     public Integer parse(String value) {
       int v = (int) Long.parseLong(value);
-      if (v < start) throw new NumberFormatException("integer too small");
-      if (v > end) throw new NumberFormatException("integer too large");
+      if (v < start) throw new NumberFormatException("integer must be at least " + start);
+      if (v > end) throw new NumberFormatException("integer must be at most " + end);
       return v;
     }
   }
