@@ -92,15 +92,21 @@ public class HdlTypes {
         contents
             .append(
                 LineBuffer.formatVhdl(
-                    "{{type}} {{1}} {{is}} {{array}} ( {{2}} {{downto}} 0 ) {{of}} std_logic_vector( ",
-                    myTypeName, myNrOfEntries))
-            .append(
-                myGenericBitWidth == null
-                    ? Integer.toString(myBitWidth - 1)
-                    : String.format("%s - 1", myGenericBitWidth))
-            .append(
-                LineBuffer.formatVhdl(
-                    " " + "{{downto}} 0);")); // Important: The leading space is required
+                    "{{type}} {{1}} {{is}} {{array}} ( {{2}} {{downto}} 0 ) {{of}} ",
+                    myTypeName, myNrOfEntries));
+        if (myGenericBitWidth == null && myBitWidth == 1) {
+          contents.append("std_logic;");
+        } else {
+          contents
+              .append("std_logic_vector( ")
+              .append(
+                  myGenericBitWidth == null
+                      ? Integer.toString(myBitWidth - 1)
+                      : String.format("%s - 1", myGenericBitWidth))
+              .append(
+                  LineBuffer.formatVhdl(
+                      " " + "{{downto}} 0);")); // Important: The leading space is required
+        }
       } else {
         contents
             .append("typedef logic [")
