@@ -27,6 +27,8 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.FocusEvent;
 import java.awt.event.FocusListener;
+import java.awt.event.KeyEvent;
+import java.awt.event.MouseEvent;
 import java.util.ArrayList;
 import java.util.EventObject;
 import java.util.LinkedList;
@@ -398,7 +400,15 @@ public class AttrTable extends JPanel implements LocaleListener {
 
     @Override
     public boolean isCellEditable(EventObject anEvent) {
-      // Asks the editor if it can start editing using anEvent.
+      if (anEvent instanceof KeyEvent) {
+        if (((KeyEvent) anEvent).getKeyCode() == KeyEvent.VK_SPACE) {
+          return true;
+        }
+        return false;
+      } 
+      if (anEvent instanceof MouseEvent) {
+        return ((MouseEvent) anEvent).getClickCount() >= 1;
+      }
       return true;
     }
 
