@@ -9,12 +9,10 @@
 
 package com.cburch.logisim.gui.main;
 
-import com.cburch.logisim.circuit.Circuit;
 import com.cburch.logisim.circuit.CircuitMutation;
 import com.cburch.logisim.circuit.Wire;
 import com.cburch.logisim.comp.Component;
 import com.cburch.logisim.comp.ComponentFactory;
-import com.cburch.logisim.comp.EndData;
 import com.cburch.logisim.data.AttributeSet;
 import com.cburch.logisim.data.Bounds;
 import com.cburch.logisim.data.Location;
@@ -28,8 +26,6 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.Iterator;
-import java.util.Map;
 import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -39,17 +35,14 @@ class SelectionBase {
   static final Logger logger = LoggerFactory.getLogger(SelectionBase.class);
   static final Set<Component> NO_COMPONENTS = Collections.emptySet();
   final HashSet<Component> selected = new HashSet<>(); // of selected
-  // Components
-  // in
-  // circuit
+  // Components in circuit
   final HashSet<Component> lifted = new HashSet<>(); // of selected
   final HashSet<Component> suppressHandles = new HashSet<>(); // of
   // Components
   final Set<Component> unionSet = CollectionUtil.createUnmodifiableSetUnion(selected, lifted);
   private final ArrayList<Selection.Listener> listeners = new ArrayList<>();
   final Project proj;
-  // Components
-  // removed
+  // Components removed
   private Bounds bounds = Bounds.EMPTY_BOUNDS;
   private boolean shouldSnap = false;
 
@@ -61,11 +54,11 @@ class SelectionBase {
     if (components.isEmpty()) {
       return Bounds.EMPTY_BOUNDS;
     } else {
-      Iterator<Component> it = components.iterator();
-      Bounds ret = it.next().getBounds();
+      final var it = components.iterator();
+      var ret = it.next().getBounds();
       while (it.hasNext()) {
-        Component comp = it.next();
-        Bounds bds = comp.getBounds();
+        final var comp = it.next();
+        final var bds = comp.getBounds();
         ret = ret.add(bds);
       }
       return ret;
@@ -73,7 +66,7 @@ class SelectionBase {
   }
 
   private static boolean shouldSnapComponent(Component comp) {
-    Boolean shouldSnapValue =
+    final var shouldSnapValue =
         (Boolean)
             comp.getFactory().getFeature(ComponentFactory.SHOULD_SNAP, comp.getAttributeSet());
     return shouldSnapValue == null || shouldSnapValue;
@@ -126,7 +119,7 @@ class SelectionBase {
   //
   private void computeShouldSnap() {
     shouldSnap = false;
-    for (Component comp : unionSet) {
+    for (final var  comp : unionSet) {
       if (shouldSnapComponent(comp)) {
         shouldSnap = true;
         return;
@@ -134,12 +127,11 @@ class SelectionBase {
     }
   }
 
-  private HashMap<Component, Component> copyComponents(
-      Collection<Component> components, boolean translate) {
+  private HashMap<Component, Component> copyComponents(Collection<Component> components, boolean translate) {
     // determine translation offset where we can legally place the clipboard
     int dx;
     int dy;
-    Bounds bds = computeBounds(components);
+    final var bds = computeBounds(components);
     for (int index = 0; ; index++) {
       // compute offset to try: We try points along successively larger
       // squares radiating outward from 0,0
@@ -178,31 +170,30 @@ class SelectionBase {
     }
   }
 
-  private HashMap<Component, Component> copyComponents(
-      Collection<Component> components, int dx, int dy, boolean translate) {
-    HashMap<Component, Component> ret = new HashMap<>();
-    for (Component comp : components) {
-      Location oldLoc = comp.getLocation();
-      AttributeSet attrs =
+  private HashMap<Component, Component> copyComponents(Collection<Component> components, int dx, int dy, boolean translate) {
+    final var ret = new HashMap<Component, Component>();
+    for (final var comp : components) {
+      final var oldLoc = comp.getLocation();
+      final var attrs =
           translate || (comp.getFactory() instanceof Rom) || (comp.getFactory() instanceof Ram)
               ? comp.getAttributeSet()
               : (AttributeSet) comp.getAttributeSet().clone();
-      int newX = oldLoc.getX() + dx;
-      int newY = oldLoc.getY() + dy;
-      Object snap = comp.getFactory().getFeature(ComponentFactory.SHOULD_SNAP, attrs);
+      var newX = oldLoc.getX() + dx;
+      var newY = oldLoc.getY() + dy;
+      final var snap = comp.getFactory().getFeature(ComponentFactory.SHOULD_SNAP, attrs);
       if (snap == null || (Boolean) snap) {
         newX = Canvas.snapXToGrid(newX);
         newY = Canvas.snapYToGrid(newY);
       }
       final var newLoc = Location.create(newX, newY, false);
-      Component copy = comp.getFactory().createComponent(newLoc, attrs);
+      final var copy = comp.getFactory().createComponent(newLoc, attrs);
       ret.put(comp, copy);
     }
     return ret;
   }
 
   void deleteAllHelper(CircuitMutation xn) {
-    for (Component comp : selected) {
+    for (final var comp : selected) {
       xn.remove(comp);
     }
     selected.clear();
@@ -219,7 +210,7 @@ class SelectionBase {
   }
 
   void duplicateHelper(CircuitMutation xn) {
-    HashSet<Component> oldSelected = new HashSet<>(selected);
+    final var oldSelected = new HashSet<Component>(selected);
     oldSelected.addAll(lifted);
     pasteHelper(xn, oldSelected);
   }
@@ -227,8 +218,8 @@ class SelectionBase {
   public void fireSelectionChanged() {
     bounds = null;
     computeShouldSnap();
-    Selection.Event e = new Selection.Event(this);
-    for (Selection.Listener l : listeners) {
+    final var e = new Selection.Event(this);
+    for (final var l : listeners) {
       l.selectionChanged(e);
     }
   }
@@ -244,12 +235,12 @@ class SelectionBase {
   }
 
   public Bounds getBounds(Graphics g) {
-    Iterator<Component> it = unionSet.iterator();
+    final var it = unionSet.iterator();
     if (it.hasNext()) {
       bounds = it.next().getBounds(g);
       while (it.hasNext()) {
-        Component comp = it.next();
-        Bounds bds = comp.getBounds(g);
+        final var comp = it.next();
+        final var bds = comp.getBounds(g);
         bounds = bounds.add(bds);
       }
     } else {
@@ -258,25 +249,25 @@ class SelectionBase {
     return bounds;
   }
 
-  private boolean hasConflictTranslated(
-      Collection<Component> components, int dx, int dy, boolean selfConflicts) {
-    Circuit circuit = proj.getCurrentCircuit();
+  private boolean hasConflictTranslated(Collection<Component> components,
+                                        int dx, int dy, boolean selfConflicts) {
+    final var circuit = proj.getCurrentCircuit();
     if (circuit == null) return false;
-    for (Component comp : components) {
+    for (final var comp : components) {
       if (!(comp instanceof Wire)) {
-        for (EndData endData : comp.getEnds()) {
+        for (final var endData : comp.getEnds()) {
           if (endData != null && endData.isExclusive()) {
-            Location endLoc = endData.getLocation().translate(dx, dy);
-            Component conflict = circuit.getExclusive(endLoc);
+            final var endLoc = endData.getLocation().translate(dx, dy);
+            final var conflict = circuit.getExclusive(endLoc);
             if (conflict != null) {
               if (selfConflicts || !components.contains(conflict)) return true;
             }
           }
         }
-        Location newLoc = comp.getLocation().translate(dx, dy);
-        Bounds newBounds = comp.getBounds().translate(dx, dy);
-        for (Component comp2 : circuit.getAllContaining(newLoc)) {
-          Bounds bds = comp2.getBounds();
+        final var newLoc = comp.getLocation().translate(dx, dy);
+        final var newBounds = comp.getBounds().translate(dx, dy);
+        for (final var comp2 : circuit.getAllContaining(newLoc)) {
+          final var bds = comp2.getBounds();
           if (bds.equals(newBounds)) {
             if (selfConflicts || !components.contains(comp2)) return true;
           }
@@ -293,7 +284,7 @@ class SelectionBase {
   void pasteHelper(CircuitMutation xn, Collection<Component> comps) {
     clear(xn);
 
-    Map<Component, Component> newLifted = copyComponents(comps, false);
+    final var newLifted = copyComponents(comps, false);
     lifted.addAll(newLifted.values());
     fireSelectionChanged();
   }
@@ -302,15 +293,15 @@ class SelectionBase {
   public void print() {
     logger.debug(" shouldSnap: {}", shouldSnap());
 
-    boolean hasPrinted = false;
-    for (Component comp : selected) {
+    var hasPrinted = false;
+    for (final var comp : selected) {
       if (hasPrinted) logger.debug("       : {}  [{}]", comp, comp.hashCode());
       else logger.debug(" select: {}  [{}]", comp, comp.hashCode());
       hasPrinted = true;
     }
 
     hasPrinted = false;
-    for (Component comp : lifted) {
+    for (final var comp : lifted) {
       if (hasPrinted) logger.debug("       : {}  [{}]", comp, comp.hashCode());
       else logger.debug(" lifted: {}  [{}]", comp, comp.hashCode());
       hasPrinted = true;
@@ -319,7 +310,7 @@ class SelectionBase {
 
   // removes from selection - NOT from circuit
   void remove(CircuitMutation xn, Component comp) {
-    boolean removed = selected.remove(comp);
+    var removed = selected.remove(comp);
     if (lifted.contains(comp)) {
       if (xn == null) {
         throw new IllegalStateException("cannot remove");
@@ -350,15 +341,15 @@ class SelectionBase {
   }
 
   void translateHelper(CircuitMutation xn, int dx, int dy) {
-    Map<Component, Component> translatedComps = copyComponents(selected, dx, dy, true);
-    for (Map.Entry<Component, Component> entry : translatedComps.entrySet()) {
+    final var translatedComps = copyComponents(selected, dx, dy, true);
+    for (final var entry : translatedComps.entrySet()) {
       xn.replace(entry.getKey(), entry.getValue());
       selected.add(entry.getValue());
     }
 
-    Map<Component, Component> liftedAfter = copyComponents(lifted, dx, dy, true);
+    final var liftedAfter = copyComponents(lifted, dx, dy, true);
     lifted.clear();
-    for (Map.Entry<Component, Component> entry : liftedAfter.entrySet()) {
+    for (final var entry : liftedAfter.entrySet()) {
       xn.add(entry.getValue());
       selected.add(entry.getValue());
     }
