@@ -34,10 +34,10 @@ public class SimulationTreeRenderer extends DefaultTreeCellRenderer {
       boolean hasFocus) {
     Component ret =
         super.getTreeCellRendererComponent(tree, value, selected, expanded, leaf, row, hasFocus);
-    SimulationTreeModel model = (SimulationTreeModel) tree.getModel();
+    final var model = (SimulationTreeModel) tree.getModel();
     if (ret instanceof JLabel label) {
       if (value instanceof SimulationTreeNode node) {
-        ComponentFactory factory = node.getComponentFactory();
+        final var factory = node.getComponentFactory();
         if (factory != null) {
           label.setIcon(new RendererIcon(factory, node.isCurrentView(model)));
         }
@@ -67,13 +67,13 @@ public class SimulationTreeRenderer extends DefaultTreeCellRenderer {
 
     @Override
     public void paintIcon(Component c, Graphics g, int x, int y) {
-      ComponentDrawContext context = new ComponentDrawContext(c, null, null, g, g);
+      final var context = new ComponentDrawContext(c, null, null, g, g);
       factory.paintIcon(context, x, y, factory.createAttributeSet());
 
       // draw magnifying glass if appropriate
       if (isCurrentView) {
-        int tx = x + 13;
-        int ty = y + 13;
+        final var tx = x + 13;
+        final var ty = y + 13;
         int[] xp = {tx - 1, x + 18, x + 20, tx + 1};
         int[] yp = {ty + 1, y + 20, y + 18, ty - 1};
         g.setColor(ProjectExplorer.MAGNIFYING_INTERIOR);
