@@ -25,8 +25,8 @@ class Toolbox extends JPanel {
   Toolbox(Project proj, Frame frame, MenuListener menu) {
     super(new BorderLayout());
 
-    ToolboxToolbarModel toolbarModel = new ToolboxToolbarModel(frame, menu);
-    Toolbar toolbar = new Toolbar(toolbarModel);
+    final var toolbarModel = new ToolboxToolbarModel(frame, menu);
+    final var toolbar = new Toolbar(toolbarModel);
     add(toolbar, BorderLayout.NORTH);
 
     toolbox = new ProjectExplorer(proj, false);
