@@ -56,6 +56,12 @@ public class AppPreferences {
   // LocalePreference
   //
   private static class LocalePreference extends PrefMonitorString {
+    /**
+     * Finds the Locale corresponding to the given language string.
+     *
+     * @param lang Language string to search for.
+     * @return The corresponding Locale or null if not found.
+     */
     private static Locale findLocale(String lang) {
       Locale[] check;
       for (int set = 0; set < 2; set++) {
@@ -74,7 +80,7 @@ public class AppPreferences {
     public LocalePreference() {
       super("locale", "");
 
-      String localeStr = this.get();
+      final var localeStr = this.get();
       if (!("".equals(localeStr))) {
         LocaleManager.setLocale(Locale.forLanguageTag(localeStr));
       }
@@ -150,7 +156,7 @@ public class AppPreferences {
   }
 
   private static File convertFile(String fileName) {
-    if (fileName == null || fileName.equals("")) {
+    if (fileName == null || fileName.isEmpty()) {
       return null;
     } else {
       final var file = new File(fileName);
@@ -171,7 +177,7 @@ public class AppPreferences {
   }
 
   private static Template getCustomTemplate() {
-    File toRead = templateFile;
+    final var toRead = templateFile;
     if (customTemplateFile == null || !(customTemplateFile.equals(toRead))) {
       if (toRead == null) {
         customTemplate = null;
@@ -408,7 +414,7 @@ public class AppPreferences {
   }
 
   public static ImageIcon getScaledImageIcon(ImageIcon icon) {
-    Image iconImage = icon.getImage();
+    final var iconImage = icon.getImage();
     return new ImageIcon(
         iconImage.getScaledInstance(getScaled(IconSize), getScaled(IconSize), Image.SCALE_SMOOTH));
   }
@@ -431,6 +437,13 @@ public class AppPreferences {
   private static MyListener myListener = null;
   private static final PropertyChangeWeakSupport propertySupport =
       new PropertyChangeWeakSupport(AppPreferences.class);
+
+  // Autosave preferences
+  public static final String AUTOSAVE_ENABLE = "autosaveEnabled";
+  public static final String AUTOSAVE_PERIOD = "autosaveInterval";
+
+  public static final PrefMonitor<Boolean> AUTOSAVE_ENABLED = create(new PrefMonitorBoolean(AUTOSAVE_ENABLE, true));
+  public static final PrefMonitor<Integer> AUTOSAVE_INTERVAL = create(new PrefMonitorInt(AUTOSAVE_PERIOD, 30));
 
   // Template preferences
   public static final int IconSize = 16;
@@ -535,6 +548,7 @@ public class AppPreferences {
   public static final int DEFAULT_COMPONENT_COLOR = 0x00000000;
   public static final int DEFAULT_COMPONENT_SECONDARY_COLOR = 0x99999999;
   public static final int DEFAULT_COMPONENT_GHOST_COLOR = 0x99999999;
+  public static final int DEFAULT_COMPONENT_ICON_COLOR = 0x00000000;
 
   // restores default grid colors
   public static void setDefaultGridColors() {
@@ -545,6 +559,7 @@ public class AppPreferences {
     COMPONENT_COLOR.set(DEFAULT_COMPONENT_COLOR);
     COMPONENT_SECONDARY_COLOR.set(DEFAULT_COMPONENT_SECONDARY_COLOR);
     COMPONENT_GHOST_COLOR.set(DEFAULT_COMPONENT_GHOST_COLOR);
+    COMPONENT_ICON_COLOR.set(DEFAULT_COMPONENT_ICON_COLOR);
   }
 
   public static final PrefMonitor<Integer> CANVAS_BG_COLOR =
@@ -561,6 +576,8 @@ public class AppPreferences {
       create(new PrefMonitorInt("componentSecondaryColor", DEFAULT_COMPONENT_SECONDARY_COLOR));
   public static final PrefMonitor<Integer> COMPONENT_GHOST_COLOR =
       create(new PrefMonitorInt("componentGhostColor", DEFAULT_COMPONENT_GHOST_COLOR));
+  public static final PrefMonitor<Integer> COMPONENT_ICON_COLOR =
+      create(new PrefMonitorInt("componentIconColor", DEFAULT_COMPONENT_ICON_COLOR));
 
 
   // Layout preferences
