@@ -34,6 +34,7 @@ public class SubtractorHdlGeneratorFactory extends AbstractHdlGeneratorFactory {
         .addWire("s_extendeddataA", EXTENDED_BITS_ID)
         .addWire("s_extendeddataB", EXTENDED_BITS_ID)
         .addWire("s_sumresult", EXTENDED_BITS_ID)
+        .addWire("n_bIn", 1)
         .addWire("s_carry", 1);
     myPorts
         .add(Port.INPUT, "dataA", NR_OF_BITS_ID, Subtractor.IN0, StdAttr.WIDTH)
@@ -63,7 +64,8 @@ public class SubtractorHdlGeneratorFactory extends AbstractHdlGeneratorFactory {
       contents.add("borrowOut <= {{not}}(s_sumresult(" + EXTENDED_BITS_STRING + "-1));");
     } else {
       contents.add("""
-          assign {s_carry,result} = dataA + ~(dataB) + ~(borrowIn);
+          assign n_bIn = ~borrowIn;
+          assign {s_carry,result} = dataA + ~(dataB) + n_bIn;
           assign borrowOut        = ~s_carry;
           """);
     }
