@@ -36,7 +36,7 @@ public class SRFlipFlop extends AbstractFlipFlop {
     @Override
     public LineBuffer getUpdateLogic() {
       return LineBuffer.getHdlBuffer()
-          .add("{{assign}} s_nextState{{=}}(s_currentState{{or}}s){{and}}{{not}}(r);");
+          .add("{{assign}} s_nextState{{=}}(s_currentState{{and}}s){{or}}({{not}}(r){{and}}s){{or}}(s_currentState{{and}}{{not}}(r));");
     }
   }
 
